Change uint32_t to uintptr_t or size_t

+++ x86/kernel/include/checks.h
@@ -4,12 +4,13 @@
 #include <x86/kernel/include/e820.h>
 
 #include <stdint.h>
+#include <stddef.h>
 #include <stdbool.h>
 
 #define CHECK_MEMORY_START      0x100000
 #define CHECK_MEMORY_END        0x800000
 
 extern bool Initial_Checks(uint32_t* mbi)__attribute__((section ("start")));
-extern bool IsMemoryAvailable(uintptr_t min, uintptr_t max, struct E820_Table_Entry* table, uint32_t size)__attribute__((section ("start")));
+extern bool IsMemoryAvailable(uintptr_t min, uintptr_t max, struct E820_Table_Entry* table, size_t size)__attribute__((section ("start")));
 
 #endif
